requirements.txt
django==2.1.7
djangorestframework==3.9.1
django-rest-auth==0.9.3
django-allauth==0.39.1
httpie
